+++ simLifeDisplayAPI/src/main/java/barsuift/simLife/j3d/tree/TreeBranchPart3D.java
@@ -21,7 +21,7 @@ package barsuift.simLife.j3d.tree;
 import java.util.List;
 
 import javax.media.j3d.Group;
-import javax.vecmath.Point3d;
+import javax.vecmath.Point3f;
 
 import barsuift.simLife.Persistent;
 
@@ -30,7 +30,7 @@ public interface TreeBranchPart3D extends Persistent<TreeBranchPart3DState> {
 
     public void addLeaf(TreeLeaf3D leaf);
 
-    public Point3d getEndPoint();
+    public Point3f getEndPoint();
 
     public List<TreeLeaf3D> getLeaves();
